Start moving the plugin around
