--- src/python/pants/backend/toml/subsystems/BUILD
@@ -1,3 +1,0 @@
-# Copyright 2023 Pants project contributors (see CONTRIBUTORS.md).
-# Licensed under the Apache License, Version 2.0 (see LICENSE).
-python_sources()
