+++ src/python/pants/backend/tools/taplo/BUILD
@@ -1,3 +1,4 @@
 # Copyright 2023 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 python_sources()
+python_tests(name="tests")
