+++ src/python/pants/backend/tools/taplo/register.py
@@ -1,3 +1,10 @@
 # Copyright 2023 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
-python_sources()
+
+from pants.backend.tools.taplo import rules as taplo_rules
+
+
+def rules():
+    return [
+        *taplo_rules.rules(),
+    ]
