+++ src/python/pants/backend/tools/taplo/rules.py
@@ -1,0 +1,74 @@
+# Copyright 2023 Pants project contributors (see CONTRIBUTORS.md).
+# Licensed under the Apache License, Version 2.0 (see LICENSE).
+
+from typing import Any
+
+from pants.backend.tools.taplo.subsystem import Taplo
+from pants.core.goals.fmt import FmtFilesRequest, FmtResult, Partitions
+from pants.core.util_rules.config_files import ConfigFiles, ConfigFilesRequest
+from pants.core.util_rules.external_tool import DownloadedExternalTool, ExternalToolRequest
+from pants.engine.fs import Digest, MergeDigests, DigestContents
+from pants.engine.platform import Platform
+from pants.engine.process import Process, ProcessResult
+from pants.engine.rules import Get, MultiGet, collect_rules, rule
+from pants.util.logging import LogLevel
+from pants.util.strutil import pluralize
+
+
+class TaploFmtRequest(FmtFilesRequest):
+    tool_subsystem = Taplo
+
+
+@rule
+async def partition_inputs(
+    request: TaploFmtRequest.PartitionRequest, taplo: Taplo
+) -> Partitions[Any]:
+    if taplo.skip:
+        return Partitions()
+
+    print(taplo)
+    return Partitions.single_partition(taplo.filter_inputs(request.files))
+
+
+@rule(desc="Format with taplo", level=LogLevel.DEBUG)
+async def taplo_fmt(request: TaploFmtRequest.Batch, taplo: Taplo, platform: Platform) -> FmtResult:
+    download_taplo_get = Get(
+        DownloadedExternalTool,
+        ExternalToolRequest,
+        taplo.get_request(platform),
+    )
+    config_files_get = Get(
+        ConfigFiles, ConfigFilesRequest, taplo.config_request(request.snapshot.dirs)
+    )
+    downloaded_taplo, config_digest = await MultiGet(download_taplo_get, config_files_get)
+    input_digest = await Get(
+        Digest,
+        MergeDigests(
+            (request.snapshot.digest, downloaded_taplo.digest, config_digest.snapshot.digest)
+        ),
+    )
+    var = await Get(DigestContents, Digest, config_digest.snapshot.digest)
+    print(var[0].content)
+    argv = [
+        downloaded_taplo.exe,
+        "fmt",
+        *taplo.args,
+        *request.files,
+    ]
+    process = Process(
+        argv=argv,
+        input_digest=input_digest,
+        output_files=request.files,
+        description=f"Run taplo on {pluralize(len(request.files), 'file')}.",
+        level=LogLevel.DEBUG,
+    )
+
+    result = await Get(ProcessResult, Process, process)
+    return await FmtResult.create(request, result)
+
+
+def rules():
+    return [
+        *collect_rules(),
+        *TaploFmtRequest.rules(),
+    ]
